Implement reading the GPIO pin for the ring indicator.
This is configured to use pin 12 on the RPi GPIO header (GPIO 18 in the SoC
labelling, but we're using board numbering).  This is an active-low input
from the RockBLOCK which goes high for 5 seconds when the network sends
a "new message" notification to us.

Remove the check_for_messages code because it's no longer needed.