fix for _read_next_line in rockblock.py
fix for method _read_next_line. It is supposed to ignore null values as well as SBDRING but right now it is just altering the order in which the values are returned. not ignoring them. This causes communications between the host system and the rockblock to get out of sync because of unexpected/out-of-order messaging (#19)